created a new DistributedTrainingExecutionParams class subclassing sdk_runnable's ExecutionParams for training job only
